Workplace Notebook and minor changes on the scene
The population age is now generated with the Rio de Janeiro scene creation.
Change on the schools network, now we are using numpy.random instead of random library, so it's consistent with the choice of library for the rest of the code.

Early alpha version of workplace generation notebook was created

Co-authored-by: Cynthia Herkenhoff <herkenhoff-cynthia@users.noreply.github.com>